refactored, renamed, and extended
The new addon name is Atemi, because this is, how I wanted to call my
PvP/arena addon from the very beginning.
Also, I refactored it to be more modular, object oriented, and added
freezing trap & ice trap to the list of (hunter) cooldowns.

The TODO list got a little update, too.